+++ src/Libraries/Nop.Core/Domain/Directory/MeasureSettings.cs
@@ -10,10 +10,10 @@ public partial class MeasureSettings : ISettings
     /// <summary>
     /// Base dimension identifier
     /// </summary>
-    public int BaseDimensionId { get; set; }
+    public long BaseDimensionId { get; set; }
 
     /// <summary>
     /// Base weight identifier
     /// </summary>
-    public int BaseWeightId { get; set; }
+    public long BaseWeightId { get; set; }
 }
