+++ src/Libraries/Nop.Core/Domain/Seo/UrlRecord.cs
@@ -8,7 +8,7 @@ public partial class UrlRecord : BaseEntity
     /// <summary>
     /// Gets or sets the entity identifier
     /// </summary>
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
 
     /// <summary>
     /// Gets or sets the entity name
@@ -28,5 +28,5 @@ public partial class UrlRecord : BaseEntity
     /// <summary>
     /// Gets or sets the language identifier
     /// </summary>
-    public int LanguageId { get; set; }
+    public long LanguageId { get; set; }
 }
