+++ src/Libraries/Nop.Data/Mapping/Builders/Catalog/ProductAttributeValueBuilder.cs
@@ -20,7 +20,7 @@ public partial class ProductAttributeValueBuilder : NopEntityBuilder<ProductAttr
         table
             .WithColumn(nameof(ProductAttributeValue.Name)).AsString(400).NotNullable()
             .WithColumn(nameof(ProductAttributeValue.ColorSquaresRgb)).AsString(100).Nullable()
-            .WithColumn(nameof(ProductAttributeValue.ProductAttributeMappingId)).AsInt32().ForeignKey<ProductAttributeMapping>();
+            .WithColumn(nameof(ProductAttributeValue.ProductAttributeMappingId)).AsInt64().ForeignKey<ProductAttributeMapping>();
     }
 
     #endregion
