+++ src/Libraries/Nop.Data/Mapping/Builders/Catalog/ProductManufacturerBuilder.cs
@@ -18,8 +18,8 @@ public partial class ProductManufacturerBuilder : NopEntityBuilder<ProductManufa
     public override void MapEntity(CreateTableExpressionBuilder table)
     {
         table
-            .WithColumn(nameof(ProductManufacturer.ManufacturerId)).AsInt32().ForeignKey<Manufacturer>()
-            .WithColumn(nameof(ProductManufacturer.ProductId)).AsInt32().ForeignKey<Product>();
+            .WithColumn(nameof(ProductManufacturer.ManufacturerId)).AsInt64().ForeignKey<Manufacturer>()
+            .WithColumn(nameof(ProductManufacturer.ProductId)).AsInt64().ForeignKey<Product>();
     }
 
     #endregion
